Move the server/API details away from the server run details
